requirements.txt
humanize==0.5.1
cx_Freeze==6.0b1
py2app==0.14
wxPython>=4.0.0b2
idna==2.5
cryptography>=2.3.0
cffi==1.10.0
pyOpenSSL==17.2.0
setuptools>=38.4.0
pyvim>=2.0.23
pyVmomi>=6.7.0
requests>=2.3.0
six>=1.7.3
gitpython>=2.0.0
matplotlib>=2.0.0
numpy>=1.12.0
